# Issues/1380 faster max count