--- a/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/AST/MaxCountPropertyShape.java
+++ b/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/AST/MaxCountPropertyShape.java
@@ -96,21 +96,27 @@ public class MaxCountPropertyShape extends PathPropertyShape {
 		PlanNode validValues = maxCountFilter.getTrueNode(BufferedPlanNode.class);
 		PlanNode invalidValues = maxCountFilter.getFalseNode(BufferedPlanNode.class);
 
-		PlanNode trimmed = new LoggingNode(new TrimTuple(validValues, 0, 1), "");
+		PlanNode mergeNode1;
+		if (!shaclSailConnection.stats.isBaseSailEmpty()) {
 
-		PlanNode unique = new LoggingNode(new Unique(trimmed), "");
+			PlanNode trimmed = new LoggingNode(new TrimTuple(validValues, 0, 1), "");
 
-		PlanNode bulkedExternalLeftOuterJoin = new LoggingNode(
-				new BulkedExternalLeftOuterJoin(unique, shaclSailConnection, path.getQuery("?a", "?c", null), false),
-				"");
+			PlanNode unique = new LoggingNode(new Unique(trimmed), "");
 
-		PlanNode groupByCount = new LoggingNode(new GroupByCount(bulkedExternalLeftOuterJoin), "");
+			PlanNode bulkedExternalLeftOuterJoin = new LoggingNode(
+					new BulkedExternalLeftOuterJoin(unique, shaclSailConnection, path.getQuery("?a", "?c", null), true),
+					"");
 
-		PlanNode directTupleFromFilter = new MaxCountFilter(groupByCount, maxCount)
-				.getFalseNode(UnBufferedPlanNode.class);
+			PlanNode groupByCount = new LoggingNode(new GroupByCount(bulkedExternalLeftOuterJoin), "");
 
-		PlanNode mergeNode1 = new UnionNode(new LoggingNode(directTupleFromFilter, ""),
-				new LoggingNode(invalidValues, ""));
+			PlanNode directTupleFromFilter = new MaxCountFilter(groupByCount, maxCount)
+					.getFalseNode(UnBufferedPlanNode.class);
+
+			mergeNode1 = new UnionNode(new LoggingNode(directTupleFromFilter, ""),
+					new LoggingNode(invalidValues, ""));
+		} else {
+			mergeNode1 = invalidValues;
+		}
 
 		if (printPlans) {
 			String planAsGraphvizDot = getPlanAsGraphvizDot(mergeNode1, shaclSailConnection);