This PR addresses GitHub issue: eclipse/rdf4j#1380 .

Briefly describe the changes proposed in this PR:

* optimize use of bulked external join
* 
*