@hmottestad 2019-04-03
Before:

```
Benchmark                                           Mode  Cnt    Score   Error  Units
MaxCountBenchmarkEmpty.noShacl                      avgt   10   32.873 ± 2.405  ms/op
MaxCountBenchmarkEmpty.shacl                        avgt   10  451.291 ± 8.769  ms/op
MaxCountBenchmarkEmpty.sparqlGroupByInsteadOfShacl  avgt   10  331.409 ± 5.078  ms/op
MaxCountBenchmarkEmpty.sparqlInsteadOfShacl         avgt   10  475.572 ± 6.308  ms/op
```


After

```
Benchmark                                           Mode  Cnt    Score   Error  Units
MaxCountBenchmarkEmpty.noShacl                      avgt   10   30.610 ± 1.424  ms/op
MaxCountBenchmarkEmpty.shacl                        avgt   10  270.945 ± 8.578  ms/op
MaxCountBenchmarkEmpty.sparqlGroupByInsteadOfShacl  avgt   10  331.519 ± 9.799  ms/op
MaxCountBenchmarkEmpty.sparqlInsteadOfShacl         avgt   10  493.683 ± 9.869  ms/op
```